\name{MyClass}
\alias{MyClass}
\docType{data}
\title{
Test Class
}
\description{
A simple test class.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("MyClass", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{species}:}{Object of class \code{"data.frame"} }
    \item{\code{scale}:}{Object of class \code{"list"} }
    \item{\code{layers}:}{Object of class \code{"character"} }
  }
}

\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(MyClass)
## maybe str(MyClass) ; plot(MyClass) ...
}
\keyword{datasets}
